refactor : MarkModal Presentation, Container 로직 분리
